requirements.txt
flask
#torch
#torchvision
numpy
requests
kubernetes
Pillow
